backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiosmtplib==3.0.2
alembic==1.14.0
databases==0.9.0
fastapi==0.115.6
passlib==1.7.4
pydantic==2.10.3
python-dotenv==1.0.1
python-jose==3.4.0
requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
psycopg2-binary==2.9.10
asyncpg==0.30.0
uvicorn==0.34.0
email_validator==2.2.0
python-multipart==0.0.18
instaloader==4.14.1
